Removes the EditableField base class and integrates its contents to LevelNameField, the only class that will actually need them.
