--- Assets/Scripts/Editing/HUD Controls/EditableField.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 88d5a8b6732ab8fe4b0b3e55d323e780
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
